Add read convenience methods
These methods differ a bit from existing ones since they don't return strings
(for which the UTF-8 encoding is known thanks to their type), but bytes without any
encoding information. But this shouldn't be too confusing, and it can be useful
e.g. for CSV.jl which expects a vector of bytes in UTF-8 rather than a string.