Improve performance
Add optimized `readbytes!` method which copies data by chunks instead of byte per byte.
Internally, avoid passing a `SubArray` to `readbytes!` as there is currently no optimized
method for them: instead, wrap the corresponding memory in an `Array`.
Increase the size of the buffer from 100 to 200 bytes, which appears to be a good tradeoff.
This makes loading a file about 10 times faster than before.